+++ bcdi/postprocessing/scripts/fit_1Dcurve.py
@@ -21,13 +21,13 @@ Fit a reciprocal space linecut over selected region using different lineshapes. 
 over all regions defined by the user, limiting the number of fitting parameters.
 """
 
-datadir = 'D:/data/P10_August2019/data/gold2_2_00515/pynx/'
+datadir = 'D:/data/P10_August2019/data/magnetite_A2_new_00013/pynx/'
 xlim = [0, 1]  # limits used for the horizontal axis of plots, leave None otherwise
 ylim = None  # [0, 3]  # limits used for the vertical axis of plots, leave None otherwise
 lineshape = 'pseudovoigt'  # lineshape to use for fitting: 'gaussian', 'lorentzian' or 'pseudovoigt'
 scale = 'log'  # scale for plots, 'linear' or 'log'
 field_names = ['distances', 'average']  # names of the fields in the file
-fit_range = [[0.40, 0.60], [0.70, 0.8]]  # list of ranges for simultaneous fit [[start1, stop1],[start2, stop2],...]
+fit_range = [[0.35, 0.65], [0.65, 0.85]]  # list of ranges for simultaneous fit [[start1, stop1],[start2, stop2],...]
 constraint_expr = []  # ['sqrt(8)/sqrt(3) * cen_1']  # list of string constraints for the fit, leave [] otherwise
 # if provided, len(constraint_expr) should be equal to len(fit_range)-1
 # sqrt(8)/sqrt(3), ratio of 220 to 111 in FCC materials